API Python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.10
python-multipart==0.0.6
pymysql==1.1.0
cryptography
langchain-google-genai